requirements.txt
scapy
setproctitle
watchdog
pyCrypto